Split up batch stuff
