+++ src/wallet/inscribe/mode.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+#[derive(PartialEq, Debug, Copy, Clone, Serialize, Deserialize, Default)]
+pub enum Mode {
+  #[serde(rename = "same-sat")]
+  SameSat,
+  #[default]
+  #[serde(rename = "separate-outputs")]
+  SeparateOutputs,
+  #[serde(rename = "shared-output")]
+  SharedOutput,
+}
